Home Navbar and Project Navbar - first upload
Created both the Navbar for the Home page and project page using the Figma designs as a template. Also created the message dropdown however that is not finalised yet. Javascript needed to make some things like usernames and profile pictures dynamic still needed.

I have added an envelope icon for the message part of my Navbar. I have also added 2 random pictures to serve as "Users" so I could test out the positioning of my profile pictures, both in the navbar, and in the message dropdown.

I pray this looks the same on everyone else's machines as it does on mine.